% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchcontrols.R
\name{matchcontrols}
\alias{matchcontrols}
\title{matchcontrols}
\usage{
matchcontrols(d, case_patientset_name = "None Given",
  controlpool_patientset_name = "None Given", controls_to_match = 1,
  match_variables = c("age", "gender", "race"))
}
\arguments{
\item{d}{A data frame of patients with matching variables and cohort field}

\item{case_patientset_name}{Name of the case patient set (for report)}

\item{controlpool_patientset_name}{Name of the control pool patient set (for report)}

\item{controls_to_match}{Number of controls to match to each patient set}

\item{match_variables}{A string vector of variable names to use for matching}
}
\value{
A list with the elements
\item{report_text}{String of the html report}
\item{html_report}{Full path to html report file in tmp directory}
\item{match_data}{The full patient data frame with match strata and which 
   controls were matched}
\item{cem_data}{The cem object from the cem function.}
\item{matched_controls}{The final list of matched controls (can be used to 
   create an i2b2 patientset)}
}
\description{
The main match controls function to match controls using cem and generate
an html report and a list of matched controls
}
\examples{
#Not run
}
