% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_matchvars.R
\name{plot_matchvars}
\alias{plot_matchvars}
\title{plot_matchvars}
\usage{
plot_matchvars(d, match_variables, grouping_variable = "cohort")
}
\arguments{
\item{d}{A data frame of patients with matching variables and cohort field}

\item{match_variables}{A string vector of variable names to use for matching}

\item{grouping_variables}{A string to use as the grouping variable.}
}
\description{
plot_matchvars
}
\examples{
#Not run
}
